requirements.txt
alembic==1.4.2
altgraph==0.17
aniso8601==8.0.0
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
biplist==1.0.3
Brotli==1.0.7
certifi==2020.6.20
click==7.1.2
dash==1.13.4
dash-core-components==1.10.1
dash-html-components==1.0.3
dash-renderer==1.5.1
dash-table==4.8.1
DateTime==4.3
distlib==0.3.1
Django==3.0.8
feedparser==5.2.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
Flask-Migrate==2.5.3
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
importlib-metadata==1.7.0
iOSbackup==0.9.9
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
macholib==1.14
Mako==1.1.3
MarkupSafe==1.1.1
modulegraph==0.18
pipenv==2020.6.2
plotly==4.9.0
py2app==0.21
pycrypto==2.6.1
PyInstaller==3.6
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
retrying==1.3.3
six==1.15.0
SQLAlchemy==1.3.18
sqlparse==0.3.1
virtualenv==20.0.28
virtualenv-clone==0.5.4
Werkzeug==1.0.1
zipp==3.1.0
zope.interface==5.1.0
